Fix: moving assets in the public folder
