+++ components/homePage1/brand.js
@@ -1,13 +1,13 @@
 import Image from "next/image";
-import b1 from "./assets/brand/b1.png";
-import b2 from "./assets/brand/b2.png";
-import b3 from "./assets/brand/b3.png";
-import b4 from "./assets/brand/b4.png";
-import b5 from "./assets/brand/b5.png";
-import b6 from "./assets/brand/b6.png";
-import b7 from "./assets/brand/b7.png";
-import b8 from "./assets/brand/b8.png";
-import b9 from "./assets/brand/b9.png";
+import b1 from "../../public/assets/homePage1/brand/b1.png";
+import b2 from "../../public/assets/homePage1/brand/b2.png";
+import b3 from "../../public/assets/homePage1/brand/b3.png";
+import b4 from "../../public/assets/homePage1/brand/b4.png";
+import b5 from "../../public/assets/homePage1/brand/b5.png";
+import b6 from "../../public/assets/homePage1/brand/b6.png";
+import b7 from "../../public/assets/homePage1/brand/b7.png";
+import b8 from "../../public/assets/homePage1/brand/b8.png";
+import b9 from "../../public/assets/homePage1/brand/b9.png";
 import Button from "../button/button";
 
 export default function Brands() {
