+++ pages/index.js
@@ -1,7 +1,7 @@
 import Button from "@/components/button/button";
 import FirstSlider from "@/components/homePage1/FirstSlider";
 import Head from "next/head";
-import peso from "../components/homePage1/assets/persone.png";
+import peso from "../public/assets/homePage1/persone.png";
 import Image from "next/image";
 import Brands from "@/components/homePage1/brand";
 import CategoryRender from "@/components/homePage1/categoryRender";
